refactor(send): split the send flow out of one 1156-line file
SendScreen.swift held every step of the external send — the address screen,
the source picker, the amount keypad, the confirm sheet — plus their shared
chrome and helpers. Each now has its own file under Send/, with the pieces two
or more steps share in Send/Components/.

Nothing changed but location. The helpers and the two chrome views were
file-scope `private`, which the split cannot preserve, so they are internal
now; SendViewModel moves in beside the screens it drives.
